requirements.txt
Django>=3.0,<4.0
djangorestframework
djoser
psycopg2-binary
dj-database-url
gunicorn
drf_yasg